+++ navigation/dependencies.gradle
@@ -25,6 +25,7 @@ ext {
     mockito                : 'org.mockito:mockito-core:2.2.27',
     hamcrest               : 'org.hamcrest:hamcrest-junit:2.0.0.0',
     commonsIO              : 'commons-io:commons-io:2.5',
+    robolectric            : 'org.robolectric:robolectric:3.3.2',
 
     // instrumentation test
     testSpoonRunner        : 'com.squareup.spoon:spoon-client:1.6.2',
